push updated project

+++ meal_management/meals/views.py
@@ -1,3 +1,4 @@
+from datetime import datetime, timedelta
 from django.shortcuts import render, redirect, get_object_or_404
 from django.db.models import Sum, Max  # Add Max here
 from .models import MealEntry, BazarEntry, Person
@@ -32,25 +33,35 @@ def meal_summary(request):
     summary_data = []
     persons = Person.objects.all()
 
-    total_meals = MealEntry.objects.all().aggregate(total_meals=Sum('meals_taken'))['total_meals'] or 0
-    total_cost = BazarEntry.objects.all().aggregate(total_cost=Sum('cost'))['total_cost'] or 0
+    
+    year = request.GET.get('year')
+    if not year:
+        year = datetime.now().year
+    month = request.GET.get('month')
+    if not month:
+        month = datetime.now().month
+    first_day = datetime(int (year), int(month), 1)
+    last_day = datetime(int (year), int(month) + 1, 1) - timedelta(days=1)
+    total_meals = MealEntry.objects.filter(date__range=(first_day, last_day)).aggregate(total_meals=Sum('meals_taken'))['total_meals'] or 0
+    total_cost = BazarEntry.objects.filter(date__range=(first_day, last_day)).aggregate(total_cost=Sum('cost'))['total_cost'] or 0
+
     person_costs = {}
     for person in persons:
-        person_cost = BazarEntry.objects.filter(person=person).aggregate(person_cost=Sum('cost'))['person_cost'] or 0
+        person_cost = BazarEntry.objects.filter(person=person).filter(date__range=(first_day, last_day)).aggregate(person_cost=Sum('cost'))['person_cost'] or 0
         person_costs[person.name] = person_cost
     
 
     cost_per_meal = total_cost / total_meals if total_meals > 0 else 0
 
     for person in persons:
-        person_meals = MealEntry.objects.filter(person=person).aggregate(person_meals=Sum('meals_taken'))['person_meals'] or 0
+        person_meals = MealEntry.objects.filter(person=person).filter(date__range=(first_day, last_day)).aggregate(person_meals=Sum('meals_taken'))['person_meals'] or 0
         person_cost = cost_per_meal * person_meals
         
 
-        total_money_given = BazarEntry.objects.filter(person=person).aggregate(total_money_given=Sum('money_given'))['total_money_given'] or 0
+        total_money_given = BazarEntry.objects.filter(person=person).filter(date__range=(first_day, last_day)).aggregate(total_money_given=Sum('money_given'))['total_money_given'] or 0
         money_balance = total_money_given - person_cost
 
-        money_left_in_cash = sum(total_money_given for total_money_given in BazarEntry.objects.values_list('money_given', flat=True)) - total_cost
+        money_left_in_cash = sum(total_money_given for total_money_given in BazarEntry.objects.filter(date__range=(first_day, last_day)).values_list('money_given', flat=True)) - total_cost
 
         data = {
             'person': person,
@@ -98,6 +109,8 @@ def person_edit(request, pk):
 
 def person_delete(request, pk):
     person = get_object_or_404(Person, pk=pk)
+    if MealEntry.objects.filter(person=person).exists():
+        return render(request, 'meals/cannot_delete_person.html', {'person': person})
     if request.method == 'POST':
         person.delete()
         return redirect('person_list')
@@ -125,3 +138,33 @@ def meal_details(request):
         'meal_data': meal_data
     }
     return render(request, 'meals/meal_details.html', context)
+
+
+def bazar_details(request):
+    bazar_entries = BazarEntry.objects.all().order_by('date')
+    
+    # Creating a dictionary to store data in the format: {date: {person: {item: {given_money: cost, money_left}}}}
+    bazar_data = {}
+    for entry in bazar_entries:
+        date = entry.date
+        person = entry.person
+        item = entry.item
+        given_money = entry.money_given
+        cost = entry.cost
+        money_left = given_money - cost
+        
+        if date not in bazar_data:
+            bazar_data[date] = {}
+        if person not in bazar_data[date]:
+            bazar_data[date][person] = {}
+        if item not in bazar_data[date][person]:
+            bazar_data[date][person][item] = {'given_money': 0, 'cost': 0, 'money_left': 0}
+        
+        bazar_data[date][person][item]['given_money'] += given_money
+        bazar_data[date][person][item]['cost'] += cost
+        bazar_data[date][person][item]['money_left'] = money_left
+        
+    context = {
+        'bazar_data': bazar_data
+    }
+    return render(request, 'meals/bazar_details.html', context)
